Ajout de la modification de LEGO
L'utilisateur connecté peut ajouter une nouvelle boite ou modifier une boite existante puis l'administrateur va pouvoir soit l'ajouter (ou modifier les infos) à la base de données ou soit le supprimer. Début des tries de console.log

+++ SAE/src/components/SuggestionBoite/SuggestionBoite.suggestionboite.js
@@ -1,4 +1,5 @@
 import SuggestionService from "../../services/SuggestionService";
+import Swal from 'sweetalert2';
 
 export default {
     props: {
@@ -6,22 +7,52 @@ export default {
     },
     methods: {
         async ajouterDansBDD() {
-            const result = await SuggestionService.addSuggestion(this.suggestion)
-            if(result.success){
-                window.location.href = "http://127.0.0.1:5173/profil";
-            }
+            Swal.fire({
+                title: 'Voulez-vous valider cette suggestion ?',
+                showDenyButton: true,
+                confirmButtonText: 'Oui',
+                denyButtonText: `Non`,
+                allowOutsideClick: false,
+                customClass: {
+                    container: 'custom-sweetalert-container',
+                    title: 'custom-sweetalert-title',
+                    content: 'custom-sweetalert-text',
+                },
+                background: 'var(--color-background)',
+            }).then(async (result) => {
+                /* Read more about isConfirmed, isDenied below */
+                if (result.isConfirmed) {
+                    const result = await SuggestionService.addSuggestion(this.suggestion)
+                    if(result.success){
+                        window.location.href = "http://127.0.0.1:5173/profil";
+                    }
+                } 
+            })
+            
         },
         async supprimerDeBDD() {
-            const result = await SuggestionService.removeSuggestion(this.suggestion)
-            if(result.success){
-                window.location.href = "http://127.0.0.1:5173/profil";
-            }
+            Swal.fire({
+                title: 'Voulez-vous vraiment supprimer cette suggestion ?',
+                showDenyButton: true,
+                confirmButtonText: 'Oui',
+                denyButtonText: `Non`,
+                allowOutsideClick: false,
+                customClass: {
+                    container: 'custom-sweetalert-container',
+                    title: 'custom-sweetalert-title',
+                    content: 'custom-sweetalert-text',
+                },
+                background: 'var(--color-background)',
+              }).then(async (result) => {
+                /* Read more about isConfirmed, isDenied below */
+                if (result.isConfirmed) {
+                    const result = await SuggestionService.removeSuggestion(this.suggestion)
+                    if(result.success){
+                        window.location.href = "http://127.0.0.1:5173/profil";
+                    }
+                }
+            })
+            
         },
-        async modifierDansBDD() {
-            const result = await SuggestionService.modifSuggestion(this.suggestion)
-            if(result.success){
-                window.location.href = "http://127.0.0.1:5173/profil";
-            }
-        }
     }
 }
